Place dataset and loss in separate modules
